Add necessary integration pieces for Travis CI
- .travis.yml: provides the metadata for describing how
  the project should be built and tested.
- travis/build.sh: builds the project.
- travis/test.sh: tests the project.

This commit adds support for the following matrix:
- Ubuntu 16.04 LTS (xenial) x {clang,gcc} x {ext4}

Support for OS X will be added once runtime issues with it
have been resolved.

Some administrative switches need to be flipped in order for
this change to be made effective, but this has proven
successful on my fork (yaneurabeya/pjdfstest).

Capture client information for FreeBSD, Linux, and OS X

FreeBSD isn't supported with Travis CI, but Linux definitely
is, and OS X is partially supported.

Fixes GitHub issue #12.