Reset ACLs between each subtest in granular/05.t
This way subtests can stand alone rather than depending on each other.

Also, fix several expectations.  Comments indicate that the correct
behavior was not observed, but rather than annotate it with a "todo"
statement the assertions were altered to expect incorrect behavior
instead.  But in fact the correct behavior is observed when starting
with a fresh ACL for each subtest.